apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
python-multipart==0.0.12
httpx==0.28.1
email-validator
bcrypt==3.2.2
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
